#11 sheet edit detection (#13)
Start working on edit detection

+++ app.py
@@ -48,13 +48,14 @@ class Service:  # pylint: disable=too-many-instance-attributes
 
         self._calendar = None
         self._gspread_client = None
-        self._sheet_dict = None
 
         self.looked_month = month
         self.looked_user = user
         self.calendar_id = ''
         self.work_days = WorkDays()
         self.event_list = []
+
+        self.last_update = 'unknown'
 
         # Get config secrets (keyfile)
         try:
@@ -80,8 +81,9 @@ class Service:  # pylint: disable=too-many-instance-attributes
             sys.exit()
 
         # TODO #4 spread sheet is user dependent
-        sheet_obj = self._gspread_client.open(year).get_worksheet(0)
-        self._sheet_dict = sheet_obj.get_all_records()
+        # Get the database (spreadsheet)
+        self.sheet_obj = self._gspread_client.open(year).get_worksheet(0)
+
         # TODO #5 validate spreadsheet format/content
         self._search_calendar_id()
         self.init_event_list_from_database()
@@ -97,9 +99,14 @@ class Service:  # pylint: disable=too-many-instance-attributes
             print(exception)
             sys.exit()
 
-    @property
-    def data(self):
-        return self._sheet_dict
+    def need_update(self) -> bool:
+        # Check update time
+        self.last_update = self.sheet_obj.spreadsheet.get_lastUpdateTime()
+        calendar_update = "not detected"  # TODO finish later #11 to manage last edition date
+        print(f"Last database change: {self.last_update}")
+        print(f"Last calendar update: {calendar_update}")
+
+        return self.last_update != calendar_update
 
     def delete_events(self, year: int, month: int):
         print("Delete old events")
@@ -158,9 +165,11 @@ class Service:  # pylint: disable=too-many-instance-attributes
         Initialize the calendar id from database
         :return:
         """
-        print("Looked event")
+        print("Get calendars from database (spreadsheet)")
 
-        for element in self.data:
+        sheet_dict = self.sheet_obj.get_all_records()
+
+        for element in sheet_dict:
             if element.get('Type') == 'Calendar':
                 # Get the calendar id
                 self.calendar_id = element.get(self.looked_user)
@@ -176,9 +185,10 @@ class Service:  # pylint: disable=too-many-instance-attributes
         Initialize the event_list content from database
         :return:
         """
-        print("Looked event")
+        print("Get events from database (spreadsheet)")
+        sheet_dict = self.sheet_obj.get_all_records()
 
-        for element in self.data:
+        for element in sheet_dict:
 
             if element.get("Date") != "":  # Event line has a non empty date column value.
                 element_date = date.fromisoformat(element.get("Date").replace("/", "-"))
@@ -339,8 +349,8 @@ def main():
         print("Use default time config")
         year = datetime.now().year
         # month = datetime.now().month
-        month = 0
-        user = 'Aurélie'  # Supported values are ['Aurélie', 'Axel', 'Aurore']
+        month = 1
+        user = 'Aurore'  # Supported values are ['Aurélie', 'Axel', 'Aurore']
 
     sheet_name = f"{year}"
 
@@ -354,14 +364,19 @@ def main():
     # colors = my.calendar.list_event_colors()
     # print(colors)
 
-    wait_before_continue(calendar_service.use_local)
-    calendar_service.delete_events(year, month)
+    if calendar_service.need_update():
 
-    wait_before_continue(calendar_service.use_local)
-    calendar_service.save_user_days()
+        wait_before_continue(calendar_service.use_local)
+        calendar_service.delete_events(year, month)
 
-    wait_before_continue(calendar_service.use_local)
-    calendar_service.print_events()
+        wait_before_continue(calendar_service.use_local)
+        calendar_service.save_user_days()
+
+        wait_before_continue(calendar_service.use_local)
+        calendar_service.print_events()
+
+    else:
+        print(f"No need to update since {calendar_service.last_update}")
 
 
 if __name__ == '__main__':
